+++ src/components/organisms/Footer/index.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import "./footer.css";
 
 export function Footer() {
-  return <footer className="nw-footer">Footer 🤔 maybe... 🤔</footer>;
+  return <footer className="nw-footer">Footer</footer>;
 }
 
 export default Footer;